fashion_prediction/src/requirements.txt
fastapi
uvicorn
python-multipart
pillow
tensorflow
numpy
streamlit